+++ example/lib/main.dart
@@ -170,6 +170,23 @@ class _MyHomePageState extends State<MyHomePage> {
     Surveys.showSurvey('PMqUZXqarkOR2yGKiENB4w');
   }
 
+  final _scaffoldKey = GlobalKey<ScaffoldState>();
+
+  void getCurrentSessionReplaylink() async {
+    final result = await SessionReplay.getSessionReplayLink();
+    if (result == null) {
+      const snackBar = SnackBar(
+        content: Text('No Link Found'),
+      );
+      ScaffoldMessenger.of(_scaffoldKey.currentContext!).showSnackBar(snackBar);
+    } else {
+      var snackBar = SnackBar(
+        content: Text(result),
+      );
+      ScaffoldMessenger.of(_scaffoldKey.currentContext!).showSnackBar(snackBar);
+    }
+  }
+
   void showFeatureRequests() {
     FeatureRequests.show();
   }
@@ -204,10 +221,11 @@ class _MyHomePageState extends State<MyHomePage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
+      key: _scaffoldKey,
       appBar: AppBar(title: Text(widget.title)),
       body: SingleChildScrollView(
           physics: ClampingScrollPhysics(),
-          padding: const EdgeInsets.only(top: 20.0),
+          padding: const EdgeInsets.only(top: 20.0, bottom: 16),
           child: Column(
             mainAxisAlignment: MainAxisAlignment.center,
             children: <Widget>[
@@ -342,7 +360,7 @@ class _MyHomePageState extends State<MyHomePage> {
               ),
               SectionTitle('Color Theme'),
               ButtonBar(
-                mainAxisSize: MainAxisSize.max,
+                mainAxisSize: MainAxisSize.min,
                 alignment: MainAxisAlignment.center,
                 children: <Widget>[
                   ElevatedButton(
@@ -363,6 +381,11 @@ class _MyHomePageState extends State<MyHomePage> {
                     child: const Text('Dark'),
                   ),
                 ],
+              ),
+              SectionTitle('Sessions Replay'),
+              InstabugButton(
+                onPressed: getCurrentSessionReplaylink,
+                text: 'Get current session replay link',
               ),
             ],
           )), // This trailing comma makes auto-formatting nicer for build methods.