+++ lib/src/modules/apm.dart
@@ -33,6 +33,11 @@ class APM {
 
   /// Starts an execution trace.
   /// [String] name of the trace.
+  ///
+  /// Please migrate to the App Flows APIs: [startFlow], [setFlowAttribute], and [endFlow].
+  @Deprecated(
+    'Please migrate to the App Flows APIs: APM.startAppFlow, APM.endFlow, and APM.setFlowAttribute. This feature was deprecated in v13.0.0',
+  )
   static Future<Trace> startExecutionTrace(String name) async {
     final id = IBGDateTime.instance.now();
     final traceId = await _host.startExecutionTrace(id.toString(), name);
@@ -54,6 +59,11 @@ class APM {
   /// [String] id of the trace.
   /// [String] key of attribute.
   /// [String] value of attribute.
+  ///
+  /// Please migrate to the App Flows APIs: [startFlow], [setFlowAttribute], and [endFlow].
+  @Deprecated(
+    'Please migrate to the App Flows APIs: APM.startAppFlow, APM.endFlow, and APM.setFlowAttribute. This feature was deprecated in v13.0.0',
+  )
   static Future<void> setExecutionTraceAttribute(
     String id,
     String key,
@@ -64,8 +74,49 @@ class APM {
 
   /// Ends an execution trace.
   /// [String] id of the trace.
+  ///
+  /// Please migrate to the App Flows APIs: [startFlow], [setFlowAttribute], and [endFlow].
+  @Deprecated(
+    'Please migrate to the App Flows APIs: APM.startAppFlow, APM.endFlow, and APM.setFlowAttribute. This feature was deprecated in v13.0.0',
+  )
   static Future<void> endExecutionTrace(String id) async {
     return _host.endExecutionTrace(id);
+  }
+
+  /// Starts an AppFlow with the given [name].
+  ///
+  /// The [name] must not be an empty string. It should be unique and not exceed 150 characters,
+  /// ignoring leading and trailing spaces.
+  ///
+  /// Duplicate [name]s will terminate the older AppFlow with the termination reason recorded as
+  /// 'force abandon end reason'.
+  ///
+  /// The method will only execute if APM is enabled, the feature is
+  /// active, and the SDK has been initialized.
+  static Future<void> startFlow(String name) async {
+    if (name.isNotEmpty) {
+      return _host.startFlow(name.trim());
+    }
+  }
+
+  /// Assigns a custom attribute to an AppFlow with the specified [name], [key], and [value].
+  ///
+  /// The [name] must not be an empty string. The [key] should not exceed 30 characters,
+  /// and [value] should not exceed 60 characters, with both ignoring leading and trailing spaces.
+  ///
+  /// To remove an attribute, set its [value] to null. Attributes cannot be added or
+  /// modified after an AppFlow has concluded.
+  static Future<void> setFlowAttribute(
+    String name,
+    String key,
+    String? value,
+  ) async {
+    return _host.setFlowAttribute(name, key, value);
+  }
+
+  /// Ends the AppFlow with the given [name].
+  static Future<void> endFlow(String name) async {
+    return _host.endFlow(name);
   }
 
   /// Enables or disables auto UI tracing.