+++ test/apm_test.dart
@@ -66,6 +66,7 @@ void main() {
     when(mHost.startExecutionTrace(id.toString(), name))
         .thenAnswer((_) async => id.toString());
 
+    // ignore: deprecated_member_use_from_same_package
     final trace = await APM.startExecutionTrace(name);
 
     expect(trace.id, id.toString());
@@ -80,6 +81,7 @@ void main() {
     const key = "attr-key";
     const attribute = "Trace Attribute";
 
+    // ignore: deprecated_member_use_from_same_package
     await APM.setExecutionTraceAttribute(id, key, attribute);
 
     verify(
@@ -90,11 +92,46 @@ void main() {
   test('[endExecutionTrace] should call host method', () async {
     final id = DateTime.now().toString();
 
+    // ignore: deprecated_member_use_from_same_package
     await APM.endExecutionTrace(id);
 
     verify(
       mHost.endExecutionTrace(id),
     ).called(1);
+  });
+
+  test('[startFlow] should call host method', () async {
+    const flowName = "flow-name";
+    await APM.startFlow(flowName);
+
+    verify(
+      mHost.startFlow(flowName),
+    ).called(1);
+    verifyNoMoreInteractions(mHost);
+  });
+
+  test('[setFlowAttribute] should call host method', () async {
+    const flowName = "flow-name";
+    const flowAttributeKey = 'attribute-key';
+    const flowAttributeValue = 'attribute-value';
+
+    await APM.setFlowAttribute(flowName, flowAttributeKey, flowAttributeValue);
+
+    verify(
+      mHost.setFlowAttribute(flowName, flowAttributeKey, flowAttributeValue),
+    ).called(1);
+    verifyNoMoreInteractions(mHost);
+  });
+
+  test('[endFlow] should call host method', () async {
+    const flowName = "flow-name";
+
+    await APM.endFlow(flowName);
+
+    verify(
+      mHost.endFlow(flowName),
+    ).called(1);
+    verifyNoMoreInteractions(mHost);
   });
 
   test('[startUITrace] should call host method', () async {